requirements.txt
openai>=1.40
pandas>=2.0
matplotlib>=3.8
pillow>=10.0
python-dotenv>=1.0